+++ ImageGallery/src/org/sleuthkit/autopsy/imagegallery/datamodel/grouping/GroupViewMode.java
@@ -1,0 +1,7 @@
+package org.sleuthkit.autopsy.imagegallery.datamodel.grouping;
+
+public enum GroupViewMode {
+
+    TILE, SLIDE_SHOW
+
+}
