--- ImageGallery/src/org/sleuthkit/autopsy/imagegallery/grouping/GroupViewMode.java
@@ -1,7 +1,0 @@
-package org.sleuthkit.autopsy.imagegallery.grouping;
-
-public enum GroupViewMode {
-
-    TILE, SLIDE_SHOW
-
-}
